(FACT-2990) Reworked virtualization structure for linux
Moved the virtual_detector resolver from the util/facts directory to the
util/linux directory and added the Linux module since without one might think
its platform-agnostic, but its actually Linux/POSIX specific. Created
virtualization folder in which we moved is_virtual and virtual facts
