+++ lib/facter/facts/linux/cloud/provider.rb
@@ -7,7 +7,7 @@ module Facts
         FACT_NAME = 'cloud.provider'
 
         def initialize
-          @virtual = Facter::Util::Facts::VirtualDetector.new
+          @virtual = Facter::Util::Facts::Posix::VirtualDetector.new
         end
 
         def call_the_resolver
