+++ spec/facter/facts/freebsd/is_virtual_spec.rb
@@ -4,10 +4,10 @@ describe Facts::Freebsd::IsVirtual do
   describe '#call_the_resolver' do
     subject(:fact) { Facts::Freebsd::IsVirtual.new }
 
-    let(:virtual_detector_double) { instance_spy(Facter::Util::Facts::VirtualDetector) }
+    let(:virtual_detector_double) { instance_spy(Facter::Util::Facts::Posix::VirtualDetector) }
 
     before do
-      allow(Facter::Util::Facts::VirtualDetector).to receive(:new).and_return(virtual_detector_double)
+      allow(Facter::Util::Facts::Posix::VirtualDetector).to receive(:new).and_return(virtual_detector_double)
       allow(virtual_detector_double).to receive(:platform).and_return(virtual_value)
     end
